M3-108 Refactor : log 삭제, 주소 업데이트 함수 분리, 코드 리펙토링
